Introduce new spring autoconfiguration support and made GraphQLEntityViewSupportFactory replacable.
