--- integration/graphql-dgs/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.blazebit.persistence.integration.graphql.dgs.BlazePersistenceDgsAutoConfiguration
