+++ integration/graphql-dgs/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,2 @@
+com.blazebit.persistence.integration.graphql.dgs.BlazePersistenceDgsAutoconfiguration
+com.blazebit.persistence.integration.graphql.dgs.BlazePersistenceSpringGraphQLConfiguration
